Add more write/read access of for save file and tests.

+++ PCPDFengineCore/Persistence/PersistenceController.cs
@@ -4,10 +4,19 @@ using PCPDFengineCore.Persistence.Records;
 
 namespace PCPDFengineCore.Persistence
 {
+    public enum AccessMode
+    {
+        READ,
+        WRITE,
+        CLOSED
+    }
+
     public class PersistenceController : IPersistenceController
     {
         private PersistenceFileDbContext? _dbContext;
         private string _dataBasePath;
+        private AccessMode _accessMode;
+        public AccessMode AccessMode { get => _accessMode; }
 
         public void UpdateFileInformation(FileInformation fileInformation)
         {
@@ -28,14 +37,44 @@ namespace PCPDFengineCore.Persistence
 
         public void LoadDatabase(string dataBasePath, bool clearExisting = false)
         {
-            _dbContext = new PersistenceFileDbContext(dataBasePath);
-
-            if (clearExisting)
+            List<Action> fileLoadMethods = new List<Action>
             {
-                _dbContext.Database.EnsureDeleted();
-            }
+                () => LoadFileWriteAccess(dataBasePath),
+                () => LoadFileReadAccess(dataBasePath)
+            };
 
-            _dbContext.Database.EnsureCreated();
+            foreach (Action loadMethod in fileLoadMethods)
+            {
+                try
+                {
+                    loadMethod();
+
+                    if (_accessMode == AccessMode.WRITE)
+                    {
+                        if (clearExisting)
+                        {
+                            _dbContext.Database.EnsureDeleted();
+                        }
+
+                        _dbContext.Database.EnsureCreated();
+                    }
+
+                    break; // We have sucessfully opened the file
+                }
+                catch (IOException ex) when (ex.Message.Contains("being used by another process"))
+                {
+                    if (_accessMode == AccessMode.READ)
+                    {
+                        _accessMode = AccessMode.CLOSED;
+                        throw; // If we still can't access the file in read mode, something is wrong.
+                    }
+                }
+                catch
+                {
+                    _accessMode = AccessMode.CLOSED;
+                    throw;
+                }
+            }
         }
 
         public void SaveDatabase()
@@ -54,6 +93,7 @@ namespace PCPDFengineCore.Persistence
         {
             if (_dbContext != null)
             {
+                _accessMode = AccessMode.CLOSED;
                 _dbContext.Database.CloseConnection();
                 _dbContext.Dispose();
                 _dbContext = null;
@@ -69,8 +109,20 @@ namespace PCPDFengineCore.Persistence
 
             return null;
         }
-
+        // Seperate value that can be toggle to public for development.
         private PersistenceFileDbContext? DbContext { get => _dbContext; }
         public string DataBasePath { get => _dataBasePath; set => _dataBasePath = value; }
+
+        public void LoadFileWriteAccess(string databasePath)
+        {
+            _dbContext = new PersistenceFileDbContext(databasePath, AccessMode.WRITE);
+            _accessMode = AccessMode.WRITE;
+        }
+
+        public void LoadFileReadAccess(string databasePath)
+        {
+            _dbContext = new PersistenceFileDbContext(databasePath, AccessMode.READ);
+            _accessMode = AccessMode.READ;
+        }
     }
 }
